CORE-17154 - standardize the keepaliveTimeSeconds - add pool config to corda.boot.json, restrict additionalProperties - remove StateManagerConfigProvider and rely on defaults from helm, apply these defaults in combined worker
