+++ gradle.properties
@@ -46,7 +46,7 @@ commonsTextVersion = 1.10.0
 bouncycastleVersion=1.73
 # Corda API libs revision (change in 4th digit indicates a breaking change)
 # Change to 5.0.0.xx-SNAPSHOT to pick up maven local published copy
-cordaApiVersion=5.1.0.20-alpha-1695130173084
+cordaApiVersion=5.1.0.20-alpha-1695215870991
 
 disruptorVersion=3.4.4
 felixConfigAdminVersion=1.9.26
