+++ libs/messaging/db-message-bus-impl/src/test/kotlin/net/corda/messagebus/db/configuration/DbBusConfigMergerImplTest.kt
@@ -30,7 +30,7 @@ class DbBusConfigMergerImplTest {
         assertThat(result.getString("stateManager.database.pass")).isEqualTo("samplepass")
         assertThat(result.getString("stateManager.database.jdbc.url")).isEqualTo("samplestatemanager")
         assertThat(result.getInt("stateManager.database.pool.idleTimeoutSeconds")).isEqualTo(120)
-        assertThat(result.getInt("stateManager.database.pool.keepAliveTimeSeconds")).isEqualTo(0)
+        assertThat(result.getInt("stateManager.database.pool.keepaliveTimeSeconds")).isEqualTo(0)
         assertThat(result.getInt("stateManager.database.pool.maxLifetimeSeconds")).isEqualTo(1800)
         assertThat(result.getInt("stateManager.database.pool.maxSize")).isEqualTo(5)
         assertThat(result.getInt("stateManager.database.pool.minSize")).isEqualTo(1)
@@ -56,7 +56,7 @@ class DbBusConfigMergerImplTest {
         assertThat(result.getString("stateManager.database.pass")).isEqualTo("samplepass")
         assertThat(result.getString("stateManager.database.jdbc.url")).isEqualTo("samplestatemanager")
         assertThat(result.getInt("stateManager.database.pool.idleTimeoutSeconds")).isEqualTo(120)
-        assertThat(result.getInt("stateManager.database.pool.keepAliveTimeSeconds")).isEqualTo(0)
+        assertThat(result.getInt("stateManager.database.pool.keepaliveTimeSeconds")).isEqualTo(0)
         assertThat(result.getInt("stateManager.database.pool.maxLifetimeSeconds")).isEqualTo(1800)
         assertThat(result.getInt("stateManager.database.pool.maxSize")).isEqualTo(5)
         assertThat(result.getInt("stateManager.database.pool.minSize")).isEqualTo(1)
