+++ libs/messaging/db-message-bus-impl/src/test/resources/test_boot.conf
@@ -30,7 +30,7 @@
 			"pass": "samplepass",
 			"pool": {
 				"idleTimeoutSeconds": 120,
-				"keepAliveTimeSeconds": 0,
+				"keepaliveTimeSeconds": 0,
 				"maxLifetimeSeconds": 1800,
 				"maxSize": 5,
 				"minSize": 1,
