+++ libs/messaging/kafka-message-bus-impl/src/test/kotlin/net/corda/messagebus/kafka/config/KafkaConfigMergerImplTest.kt
@@ -29,7 +29,7 @@ class KafkaConfigMergerImplTest {
         assertThat(result.getString("stateManager.database.pass")).isEqualTo("samplepass")
         assertThat(result.getString("stateManager.database.jdbc.url")).isEqualTo("samplestatemanager")
         assertThat(result.getInt("stateManager.database.pool.idleTimeoutSeconds")).isEqualTo(120)
-        assertThat(result.getInt("stateManager.database.pool.keepAliveTimeSeconds")).isEqualTo(0)
+        assertThat(result.getInt("stateManager.database.pool.keepaliveTimeSeconds")).isEqualTo(0)
         assertThat(result.getInt("stateManager.database.pool.maxLifetimeSeconds")).isEqualTo(1800)
         assertThat(result.getInt("stateManager.database.pool.maxSize")).isEqualTo(5)
         assertThat(result.getInt("stateManager.database.pool.minSize")).isEqualTo(1)
@@ -53,7 +53,7 @@ class KafkaConfigMergerImplTest {
         assertThat(result.getString("stateManager.database.pass")).isEqualTo("samplepass")
         assertThat(result.getString("stateManager.database.jdbc.url")).isEqualTo("samplestatemanager")
         assertThat(result.getInt("stateManager.database.pool.idleTimeoutSeconds")).isEqualTo(120)
-        assertThat(result.getInt("stateManager.database.pool.keepAliveTimeSeconds")).isEqualTo(0)
+        assertThat(result.getInt("stateManager.database.pool.keepaliveTimeSeconds")).isEqualTo(0)
         assertThat(result.getInt("stateManager.database.pool.maxLifetimeSeconds")).isEqualTo(1800)
         assertThat(result.getInt("stateManager.database.pool.maxSize")).isEqualTo(5)
         assertThat(result.getInt("stateManager.database.pool.minSize")).isEqualTo(1)
